pyproject.toml
[tool.poetry]
name = "mapaly"
version = "0.1.0"
description = ""
authors = ["Adrián Arroyo Calle <adrian.arroyocalle@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
djangorestframework = "^3.13.0"
uwsgi = "^2.0.18"
azure-storage-blob = "^12.5.0"
lunr = {extras = ["languages"], version = "^0.6.2"}
sentry-sdk = "^1.9.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
selenium = "^3.141.0"
SPARQLWrapper = "^1.8.5"
aiohttp = "^3.7.4"

[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
black = "^22.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.mypy]
ignore_missing_imports = true